requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
